requirements.txt
googleapis-common-protos==1.5.6
google~=2.0.1
grpcio~=1.20.0
boto3~=1.9.197
pyarrow~=0.14.1

hydro-serving-grpc
requests~=2.22.0